# static/dist/api.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: c't Task API
  license:
    name: MIT
servers:
  - url: "http://server/cttaskapi/"
paths:
  /tasks:
    get:
      summary: Get all tasks
      operationId: getTasks
      tags:
        - tasks
      responses:
        "200":
          description: An array of tasks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tasks"
    post:
      summary: Create a task
      operationId: createTask
      tags:
        - tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Task"
      responses:
        "201":
          description: Null response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{taskId}:
    get:
      summary: Info for a specific Task
      operationId: getTaskById
      tags:
        - tasks
      parameters:
        - name: taskId
          in: path
          required: true
          description: The id of the Task to retrieve
          schema:
            type: string
      responses:
        "200":
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    put:
      summary: update a specific task
      operationId: updateTask
      tags:
        - tasks
      parameters:
        - name: taskId
          in: path
          required: true
          description: id of the task to update
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Task"
      responses:
        "201":
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      summary: delete a task
      operationId: deleteTask
      tags:
        - tasks
      parameters:
        - name: taskId
          in: path
          required: true
          description: id of the task to delete
          schema:
            type: string
      responses:
        "204":
          description: expected response to a valid request

components:
  schemas:
    Task:
      type: object
      required:
        - id
        - text
        - done
      properties:
        id:
          type: integer
          readOnly: true
        text:
          type: string
          minLength: 2
          maxLength: 255
        done:
          type: boolean
    Tasks:
      type: array
      items:
        $ref: "#/components/schemas/Task"